requirements.txt
audiofile==1.1.1
audioread==3.0.0
audobject==0.7.7
audresample==1.1.1
librosa==0.9.1
llvmlite==0.31.0
matplotlib==3.6.2
numba==0.48.0
numpy==1.20.3
opensmile==2.4.2
pandas==1.5.2
requests==2.28.1
resampy==0.3.1
scikit-learn==1.2.0
scipy==1.10.0
soundfile==0.11.0
torch-summary==1.4.4
tqdm==4.64.1
urllib3==1.26.13